HBASE-21476 Support date-tiered compactions when using nanosecond timestamps
This extend the original nanosecond patch such that data-tiered works
correclty when the table use nanosecond timestamps. Without this patch,
the HFile are not grouped correclty in multiple windows of time and thus
compaction creates a single window that contains all the data.

The main idea of this patch is to convert the nanosecond to millisecond.
This make things simpler as the DateTieredCompactionPolicy compares file
modification timestamp and cell timestamp, as well as all the
configuration being in millisecond.