HBASE-28488 Use encoded name in region span attributes
On our busy clusters, the alloc profile shows that createRegionSpan() is
responsible for 15-20% of all the allocations. These allocations comes
from getRegionNameAsString().

getRegionNameAsString() takes the region name and encode invisible
characters in their hex representation. This requires the use of a
StringBuilder and thus generate new strings every time.

This becomes really expensive on a cluster with high number of requests.
It seems better to just take the encoded region name (the md5 part) and
use that in trace attributes, because:
- it's fixed in size (the full region name can be much longer depending
  on the rowkey size),
- it's enough information to link a trace to a region,
- it doesn't require any new allocation.

Signed-off-by: Thibault Deutsch <thibault@arista.com>