HBASE-28569: fix race condition during WAL splitting leading to corrupt recovered.edits
If an exception happens in the call to finishWriterThreads in the
org.apache.hadoop.hbase.wal.RecoveredEditsOutputSink.close method, the call to closeWriters
should not execute, as it may lead to a race condition that leads to file corruption if
the regionserver aborts. The execution of closeWriters in this case would write the
trailer in parallel with writer threads, causing corruption, and then the corrupt file
would get renamed and finalized when it should not be. This corruption causes problems when
the region is then to be assigned. By removing the try finally block, the problematic
closeWriters would not execute in the case of an exception in finishWriterThreads, which
should then prevent this race from occurring and causing recovered.edits corruption.